+++ Changes.md
@@ -8,7 +8,11 @@ Features
 - ColorChooser :
   - Added sliders for TMI (temperature, magenta, intensity) color space.
   - Added color field, a widget giving control of two channels of "RGB", "HSV" or "TMI" triplets. The third channel is held constant.
-  - The color field, HSV slider group and TMI slider group can now be toggled on or off.
+  - The color field, HSV slider group and TMI slider group can now be toggled on or off. Visibility of the elements is maintained for the duration of the Gaffer session.
+  - Default visibility of the UI elements can now be set at startup by setting `colorChooser:inlineOptions` and `colorChooser:dialogueOptions` `userData` entries for the inline chooser and dialogue chooser respectively. The value of the metadata entry is an `IECore.CompoundData` object with the following optional keys :
+    - `visibleComponents` : A string where each character is a visible component slider. Optional components are `rgbahsvtmi`.
+    - `staticComponent` : A single character string for the component to use as the static component for the color field.
+    - `colorFieldVisible` : A boolean indicating if the color field should be visible or not.
 
 Improvements
 ------------